requirements-gpu.txt
absl-py==2.3.1
annotated-types==0.7.0
asttokens==3.0.1
astunparse==1.6.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
docstring_parser==0.17.0
exceptiongroup==1.3.1
executing==2.2.1
Flask==3.1.2
flask-cors==6.0.1
flask-swagger-ui==5.21.0
flatbuffers==25.9.23
fonttools==4.61.0
gast==0.4.0
google-api-core==2.28.1
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.5.0
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.63.1
grpc-google-iam-v1==0.13.0
grpcio==1.76.0
grpcio-status==1.48.2
h5py==3.15.1
idna==3.11
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
pillow_heif==1.1.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==3.19.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
stack-data==0.6.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==2.0.1
